Bug fixed

+++ planIt/src/main/java/planit/project/controllers/ChatController.java
@@ -64,7 +64,7 @@ public class ChatController {
 	}
 
 	@GetMapping("/all")
-	public ResponseEntity<List<MessageDTO>> checkUser(@RequestParam Integer teamId) {
+	public ResponseEntity<List<MessageDTO>> checkUser(@RequestParam Long teamId) {
 		List<MessageDTO> messages = chatService.getMessages(teamId);
 		if (messages != null) {
 			return new ResponseEntity<>(messages, HttpStatus.OK);
